Move files from coreSDK to channel and common
