+++ AppInsightsCommon/Interfaces/Telemetry/IEnvelope.ts
