+++ AppInsightsCommon/Interfaces/Telemetry/ISerializable.ts
