+++ AppInsightsCommon/Util.ts
@@ -2,6 +2,7 @@
 import { _InternalLogging } from "./Logging";
 import { IConfig } from "./Interfaces/IConfig";
 import { RequestHeaders } from "./RequestResponseHeaders";
+import { DataSanitizer } from "./Telemetry/Common/DataSanitizer";
 
 export class Util {
     private static document: any = typeof document !== "undefined" ? document : {};
@@ -644,5 +645,33 @@ export class CorrelationIdHelper {
                 }
             }
         }
+    }
+}
+
+export class AjaxHelper {
+    public static ParseDependencyPath(absoluteUrl: string, method: string, pathName: string) {
+        var target, name;
+        if (absoluteUrl && absoluteUrl.length > 0) {
+            var parsedUrl: HTMLAnchorElement = UrlHelper.parseUrl(absoluteUrl)
+            target = parsedUrl.host;
+            if (parsedUrl.pathname != null) {
+                var pathName: string = (parsedUrl.pathname.length === 0) ? "/" : parsedUrl.pathname;
+                if (pathName.charAt(0) !== '/') {
+                    pathName = "/" + pathName;
+                }
+
+                name = DataSanitizer.sanitizeString(method ? method + " " + pathName : pathName);
+            } else {
+                name = DataSanitizer.sanitizeString(absoluteUrl);
+            }
+        } else {
+            target = pathName;
+            name = pathName;
+        }
+
+        return {
+            target: target,
+            name: name
+        };
     }
 }
