+++ applicationinsights-common.ts
@@ -1,6 +1,19 @@
 export { Util, CorrelationIdHelper, UrlHelper } from './AppInsightsCommon/Util';
-export { _InternalMessageId, LoggingSeverity } from './AppInsightsCommon/Enums';
+export { _InternalMessageId, LoggingSeverity, FieldType } from './AppInsightsCommon/Enums';
 export { IConfig } from './AppInsightsCommon/Interfaces/IConfig';
 export { _InternalLogging, _InternalLogMessage } from './AppInsightsCommon/Logging';
 export { RequestHeaders } from './AppInsightsCommon/RequestResponseHeaders';
 export { DisabledPropertyName } from './AppInsightsCommon/Constants';
+export { Data as AIData } from './AppInsightsCommon/Interfaces/Contracts/Generated/Data';
+export { Base as AIBase } from './AppInsightsCommon/Interfaces/Contracts/Generated/Base';
+export { ISerializable } from './AppInsightsCommon/Interfaces/Telemetry/ISerializable';
+export { IEnvelope } from './AppInsightsCommon/Interfaces/Telemetry/IEnvelope';
+export { Envelope } from './AppInsightsCommon/Telemetry/Common/Envelope';
+export { Event } from './AppInsightsCommon/Telemetry/Event';
+export { Exception } from './AppInsightsCommon/Telemetry/Exception';
+export { Metric } from './AppInsightsCommon/Telemetry/Metric';
+export { PageView } from './AppInsightsCommon/Telemetry/PageView';
+export { RemoteDependencyData } from './AppInsightsCommon/Telemetry/RemoteDependencyData';
+export { Trace } from './AppInsightsCommon/Telemetry/Trace';
+export { PageViewPerformance } from './AppInsightsCommon/Telemetry/PageViewPerformance';
+export { Data } from './AppInsightsCommon/Telemetry/Common/Data';
