--- coreSDK/JavaScriptSDK/Telemetry/Common/Base.ts
@@ -1,10 +1,0 @@
-﻿import { Base as AIBase } from '../../../JavaScriptSDK.Interfaces/Contracts/Generated/Base';
-import { ISerializable } from '../../../JavaScriptSDK.Interfaces/Telemetry/ISerializable';
-
-export class Base extends AIBase implements ISerializable {
-
-    /**
-     * The data contract for serializing this object.
-     */
-    public aiDataContract = {};
-}
