+++ coreSDK/JavaScriptSDK/ajax/ajaxUtils.ts
@@ -87,31 +87,3 @@ export class EventHelper {
         }
     }
 }
-
-export class AjaxHelper {
-    public static ParseDependencyPath(absoluteUrl: string, method: string, pathName: string) {
-        var target, name;
-        if (absoluteUrl && absoluteUrl.length > 0) {
-            var parsedUrl: HTMLAnchorElement = UrlHelper.parseUrl(absoluteUrl)
-            target = parsedUrl.host;
-            if (parsedUrl.pathname != null) {
-                var pathName: string = (parsedUrl.pathname.length === 0) ? "/" : parsedUrl.pathname;
-                if (pathName.charAt(0) !== '/') {
-                    pathName = "/" + pathName;
-                }
-
-                name = DataSanitizer.sanitizeString(method ? method + " " + pathName : pathName);
-            } else {
-                name = DataSanitizer.sanitizeString(absoluteUrl);
-            }
-        } else {
-            target = pathName;
-            name = pathName;
-        }
-
-        return {
-            target: target,
-            name: name
-        };
-    }
-}
