+++ coreSDK/amd/package.json
@@ -2,7 +2,8 @@
     "name": "applicationinsights-core-js",
     "version": "0.0.1",
     "description": "Microsoft Application Insights Core Javascript SDK",
-    "main": "bundle/aicore.js",
+    "main": "bundle/applicationinsights-core-js.js",
+    "types": "bundle/applicationinsights-core-js.d.ts",
     "devDependencies": {
         "grunt": "1.0.1",
         "grunt-contrib-qunit": "2.0.0",
