+++ coreSDK/cjs/package.json
@@ -9,6 +9,6 @@
         "grunt-ts": "^6.0.0-beta.15",
         "typescript": "2.5.3"
     },
-    "main": "coreSDK/applicationinsights-core-js.js",
-    "types": "coreSDK/applicationinsights-core-js.d.ts"
+    "main": "bundle/applicationinsights-core-js.js",
+    "types": "bundle/applicationinsights-core-js.d.ts"
 }
